Allow name as a tag for exemplars
name was being used as a keyword to store the counter/gauge name for exemplars, so if users had a tag of name for one of these objects, that tag would be overwritten in storage and lookup. Changed keyword to <<"__name"__>> instead to avoid overriding user-defined tags as much as possible.