Update counters and exemplars to follow OpenMetrics
According to the OpenMetrics standard, counters must have the suffix "_total" appended to them, and any counters that don't do so are considered to be of "unknown" type instead. Additionally, exemplars are only supported on limited data types, and within our code base, are only officially supported on counters. The code has been updated to reflect this change.