Limit bucket count, fix hist range constructor
Added an environment variable to limit the maximum number of buckets allowed in a histogram as a failsafe if people unintentionally create massive histograms, mostly useful to prevent cost blowups as each bucket is considered a separate time series. Additionally, discovered and fixed a bug with the range based constructor for histograms where 1. the min bucket would sometimes not be created due to floating point errors, and 2. more buckets would sometimes be generated than intended. Added/modified tests to account for both the new feature and the bug fix.