Support OpenMetrics Info, set_exemplar Histogram Support
Add set_info/2, as well as related structure in existing functions, to support OpenMetrics info metrics with unique names.
Modified imetrics:set_exemplar to have more function variants to account for a Type argument, which assumes counter but can be specified to be histogram as an alternative to calling imetrics_hist_openmetrics:set_exemplar. Currently supported types are histogram and counter, as limited by OpenMetrics.